+++ src/Datasets/Labeled.php
@@ -113,6 +113,30 @@ class Labeled extends Dataset
     }
 
     /**
+     * Build a dataset from a newline delimited JSON string.
+     *
+     * @param string $ndjson
+     * @return self
+     */
+    public static function fromNdjson(string $ndjson) : self
+    {
+        $rows = preg_split(self::NEWLINE_REGEX, $ndjson) ?: [];
+
+        $samples = $labels = [];
+
+        foreach ($rows as $row) {
+            if (!empty($row)) {
+                $data = json_decode($row, true);
+
+                $samples[] = array_slice($data, 0, -1);
+                $labels[] = end($data);
+            }
+        }
+
+        return self::build($samples, $labels);
+    }
+
+    /**
      * Build a dataset object from a CSV string.
      *
      * @param string $csv
@@ -1094,6 +1118,22 @@ class Labeled extends Dataset
     public function toJson(bool $pretty = false) : string
     {
         return json_encode($this, $pretty ? JSON_PRETTY_PRINT : 0) ?: '';
+    }
+
+    /**
+     * Return a newline delimited JSON representation of the dataset.
+     *
+     * @return string
+     */
+    public function toNdjson() : string
+    {
+        $ndjson = '';
+
+        foreach ($this->zip() as $row) {
+            $ndjson .= json_encode($row) . PHP_EOL;
+        }
+
+        return $ndjson;
     }
 
     /**