+++ CHANGELOG.md
@@ -9,7 +9,8 @@
     - Added label type validation for supervised learners
     - Added toArray, toJson, and toCsv methods to Dataset API
     - Can now preview a Dataset object in console by echoing it
-    - Added JSON and CSV factories to Dataset API
+    - Added JSON, and CSV factories to Dataset API
+    - Added NDJSON support to Dataset objects
 
 - 0.0.17-beta
     - Added Tensor extension compatibility